# app/docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: PromptHub Router API
  version: 0.1.0
  description: |
    Centralized MCP router with server management, prompt enhancement, and caching.

    ## Features
    - MCP server lifecycle management (install, start, stop, monitor)
    - Unified MCP server access via JSON-RPC proxy
    - Prompt enhancement via Ollama with per-client routing
    - Response caching (L1 memory, L2 semantic)
    - Circuit breaker resilience patterns
    - Real-time monitoring dashboard
    - Audit logging and security alerts

  contact:
    name: PromptHub Project
    url: https://github.com/yourusername/prompthub
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:9090
    description: Local development server
...
paths:
...
  /servers:
    get:
      summary: List all MCP servers
      description: Get list of all configured MCP servers with their current status
      operationId: listServers
      tags:
        - Servers
      responses:
        '200':
          description: List of servers
          content:
            application/json:
              schema:
                type: object
                properties:
                  servers:
                    type: array
                    items:
                      $ref: '#/components/schemas/ServerInfo'
        '503':
          description: Server registry not initialized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /servers/{name}:
    get:
      summary: Get server details
      description: Get detailed configuration and status for a specific MCP server
      operationId: getServer
      tags:
        - Servers
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Server details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerDetail'
        '404':
          description: Server not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      summary: Remove server
      description: Remove a server configuration (must be stopped first)
      operationId: removeServer
      tags:
        - Servers
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Server removed
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Server fetch removed"
        '400':
          description: Cannot remove running server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Server not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    ServerInfo:
      type: object
      properties:
        name:
          type: string
        package:
          type: string
        transport:
          type: string
          enum: [stdio, sse, http]
        status:
          type: string
        pid:
          type: integer
          nullable: true
        auto_start:
          type: boolean
        restart_count:
          type: integer
        description:
          type: string
    ServerDetail:
      type: object
      properties:
        name:
          type: string
        config:
          type: object
        process:
          type: object
...
    Error:
      type: object
      properties:
        detail:
          type: string